refactor: Align all models and DAOs with PostgreSQL schema
- Refactor all model classes to use Lombok annotations (@Data, @NoArgsConstructor, @AllArgsConstructor)
- Update field types to match PostgreSQL schema exactly (Long for IDs, BigDecimal for monetary fields)
- Add missing model classes: Order, OrderItem, Promotion, Status, Review, ReviewLike, Viewed
- Fix all DAO classes to use correct getter/setter method names
- Update type mappings in DAO classes (int -> Long, double -> BigDecimal)
- Fix service and controller classes to use refactored models
- Ensure Maven build passes successfully
- Remove boilerplate code with Lombok, improving maintainability

+++ src/main/java/controller/CartServlet.java
@@ -17,8 +17,8 @@ public class CartServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        User user = (User) req.getSession().getAttribute("user"); // giả sử login đã có
-        int userId = user.getId();
+    User user = (User) req.getSession().getAttribute("user"); // giả sử login đã có
+    int userId = user.getId().intValue();
         try {
             List<CartItem> cart = cartDAO.getCartByUser(userId);
             req.setAttribute("cart", cart);
